% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocodebook-package.R
\docType{package}
\name{autocodebook-package}
\alias{autocodebook}
\alias{autocodebook-package}
\title{autocodebook: Automatic Codebook and Tracking for Spark/dplyr Pipelines}
\description{
Wraps dplyr verbs to automatically capture variable metadata (type, source
columns, categories, and source code), producing a codebook and an
eligibility tracking table with zero manual documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patriciafortesm/autocodebook}
  \item Report bugs at \url{https://github.com/patriciafortesm/autocodebook/issues}
}

}
\author{
\strong{Maintainer}: Patricia Fortes C. de Macedo \email{macedopatriciafortes@gmail.com}

Authors:
\itemize{
  \item Patricia Fortes C. de Macedo \email{macedopatriciafortes@gmail.com}
}

}
\keyword{internal}
