% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_render.R
\name{cb_render}
\alias{cb_render}
\title{Render the codebook as a gt table}
\usage{
cb_render(group_by_block = TRUE, show_code = TRUE)
}
\arguments{
\item{group_by_block}{Logical. If TRUE and blocks are defined,
groups rows by block. Default: TRUE.}

\item{show_code}{Logical. Show the "code" column? Default: TRUE.}
}
\value{
A gt object.
}
\description{
Render the codebook as a gt table
}
