% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_state.R
\name{cb_set_default_cache}
\alias{cb_set_default_cache}
\title{Toggle default caching for big-data verbs}
\usage{
cb_set_default_cache(default_cache = TRUE)
}
\arguments{
\item{default_cache}{Logical.}
}
\value{
Invisible previous value.
}
\description{
Toggle default caching for big-data verbs
}
