% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_state.R
\name{cb_set_verbose}
\alias{cb_set_verbose}
\title{Toggle verbose diagnostic messages}
\usage{
cb_set_verbose(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical.}
}
\value{
Invisible previous value.
}
\description{
Controls whether track_step(), auto_filter() and cb_checkpoint() print
diagnostic messages (n removed, elapsed time, etc.).
}
