% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_flow.R
\name{flow_get}
\alias{flow_get}
\title{Get the current flow tree (raw structure)}
\usage{
flow_get()
}
\value{
A list describing the flow tree.
}
\description{
Returns the internal flow representation. Mostly for debugging /
programmatic access. For a tidy table, use \code{flow_table()}.
}
