% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_render.R
\name{track_render}
\alias{track_render}
\title{Render the tracking log as a gt table}
\usage{
track_render(show_elapsed = FALSE)
}
\arguments{
\item{show_elapsed}{Logical. Show the elapsed_s column if present?
Default: FALSE.}
}
\value{
A gt object.
}
\description{
Render the tracking log as a gt table
}
